extracted out of topic folder
no longer needed that we have a single file for crash reporting
