+++ packages/core/src/shared/crashMonitoring.ts
@@ -3,19 +3,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import path from 'path'
-import globals, { isWeb } from '../extensionGlobals'
-import { getSessionId as _getSessionId } from '../telemetry/util'
-import { getErrorId, getTelemetryReason, getTelemetryReasonDesc, isFileNotFoundError, ToolkitError } from '../errors'
-import { isAutomation, isDebugInstance } from '../vscode/env'
-import { DevSettings } from '../settings'
+import globals, { isWeb } from './extensionGlobals'
+import { getSessionId as _getSessionId } from './telemetry/util'
+import { getErrorId, getTelemetryReason, getTelemetryReasonDesc, isFileNotFoundError, ToolkitError } from './errors'
+import { isAutomation, isDebugInstance } from './vscode/env'
+import { DevSettings } from './settings'
 import vscode from 'vscode'
-import { telemetry } from '../telemetry'
-import { Logger } from '../logger'
-import { isNewOsSession } from '../utilities/osUtils'
+import { telemetry } from './telemetry'
+import { Logger } from './logger'
+import { isNewOsSession } from './utilities/osUtils'
 import nodeFs from 'fs/promises'
-import fs from '../fs/fs'
-import { getLogger } from '../logger/logger'
-import { crashMonitoringDirName } from '../constants'
+import fs from './fs/fs'
+import { getLogger } from './logger/logger'
+import { crashMonitoringDirName } from './constants'
 
 const className = 'CrashMonitoring'
