+++ packages/core/src/shared/index.ts
@@ -51,4 +51,4 @@ export * as errors from './errors'
 export * as funcUtil from './utilities/functionUtils'
 export { fs } from './fs/fs'
 export * from './handleUninstall'
-export { CrashMonitoring } from './crashMonitoring/crashMonitoring'
+export { CrashMonitoring } from './crashMonitoring'
