+++ packages/core/src/testInteg/shared/crashMonitoring/crashMonitoring.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { crashMonitoringTest } from '../../../test/shared/crashMonitoring/crashMonitoring.test'
+import { crashMonitoringTest } from '../../../test/shared/crashMonitoring.test'
 
 // This test is slower so we want to it to run as an integ test for CI
 describe('CrashMonitoring', crashMonitoringTest)
